+++ packages/transaction-pay-controller/src/strategy/relay/types.ts
@@ -73,6 +73,7 @@ export type RelayQuote = {
   };
   metamask: {
     gasLimits: number[];
+    isMaxGasStation?: boolean;
   };
   request: RelayQuoteRequest;
   steps: {